Increase backlog, supply new req object after fail
If we failed to accept a connection don't retry with the same
request object again. Supply a new one.
In a period with many requests the backlog size of 5 seemed not
sufficient and lead to the case where we were unable to accept
a connection, so increase the backlog size to have some more
room.